requirements_dev.txt
pip==22.2.2
bump2version==1.0.1
wheel==0.37.1
watchdog==2.1.9
flake8==5.0.4
tox==3.26.0
coverage==6.5.0
Sphinx==4.5.0
nbsphinx==0.8.9
ipython==8.5.0
sphinx-book-theme==0.3.3
twine==4.0.1
docformatter==1.5.0
ipywidgets==8.0.7

pytest==7.1.3
black==22.10.0
mypy==0.982

bokeh==2.4.3
loguru==0.6.0
numpy==1.23.1
pytest==7.1.3
setuptools==65.3.0
terra-sdk==2.0.6
terra-proto==1.0.1
python-binance==1.0.17
pandas==2.0.3
scipy==1.11.1
statsmodels==0.14.0
tqdm==4.65.0
joblib==1.2.0
